Ajuste a solicitudes del cliente después de la primera prueba.
Se permite búsqueda de cliente por nombre
Se permite el ingreso de descuentos.

+++ lib/src/model/registro_item_model.dart
@@ -28,7 +28,7 @@ class RegistroItemModel {
   String vendedor;
   String articulo;
   int cantidad;
-  int descuento;
+  double descuento;
   bool esnumerado;
   bool sobrestock;
   String fecha;
